+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to `httpclient` will be documented in this file.
 
 ## [Unreleased]
 
-- Added support to configure global default request options, via static methods: `defaultOptions`, `setDefaultOptions`
+- Added support to configure global default request options, via static method `setDefaultOptions`
 - `option` method can accept array
 - Rename `withExceptions()` to `catchExceptions()`, added `areExceptionsCaught()`